+++ reproject/hips/__init__.py
@@ -1,1 +1,1 @@
-from .core import *  # noqa
+from .high_level import *  # noqa
